+++ src/unitycrunch/crn_unpacker.rs
@@ -338,14 +338,14 @@ impl<'slice> CrnUnpacker<'slice>{
                 let mut d = (3 * (p + 1)) as i32;
                 let mut byte_offset = d >> 3;
                 let mut bit_offset = d & 7;
-                WRITE_OR_U8_INTO_U16_BUFFER!(self.m_alpha_selectors, data_pos + byte_offset as usize, (((s as u16) << (8 - bit_offset)) as u16 & 0xFF));
+                WRITE_OR_U8_INTO_U16_BUFFER!(self.m_alpha_selectors, data_pos + byte_offset as usize, (((s as u16) << (8 - bit_offset)) & 0xFF));
                 if bit_offset < 3 {
                     WRITE_OR_U8_INTO_U16_BUFFER!(self.m_alpha_selectors, data_pos + (byte_offset as usize) - 1, (s >> bit_offset) as u16);
                 }
                 d += 9 * ((p as i32 & 3) - (p as i32 >> 2));
                 byte_offset = d >> 3;
                 bit_offset = d & 7;
-                WRITE_OR_U8_INTO_U16_BUFFER!(self.m_alpha_selectors, data_pos + (byte_offset as usize) + 6, ((s as u16) << (8 - bit_offset)) as u16 & 0xFF);
+                WRITE_OR_U8_INTO_U16_BUFFER!(self.m_alpha_selectors, data_pos + (byte_offset as usize) + 6, ((s as u16) << (8 - bit_offset)) & 0xFF);
                 if bit_offset < 3 {
                     WRITE_OR_U8_INTO_U16_BUFFER!(self.m_alpha_selectors, data_pos + (byte_offset as usize) + 5, (s >> bit_offset) as u16);
                 }
@@ -387,7 +387,7 @@ impl<'slice> CrnUnpacker<'slice>{
                 let d = (3 * ((p as i32) + 1) + 9 * (((p as i32) & 3) - ((p as i32) >> 2))) as i16;
                 let byte_offset = d >> 3;
                 let bit_offset = d & 7;
-                WRITE_OR_U8_INTO_U16_BUFFER!(self.m_alpha_selectors, i + byte_offset as usize, (((s as u16) << (8 - bit_offset)) as u16 & 0xFF));
+                WRITE_OR_U8_INTO_U16_BUFFER!(self.m_alpha_selectors, i + byte_offset as usize, (((s as u16) << (8 - bit_offset)) & 0xFF));
                 if bit_offset < 3 {
                     WRITE_OR_U8_INTO_U16_BUFFER!(self.m_alpha_selectors, i + (byte_offset as usize) - 1, (s >> bit_offset) as u16);
                 }
@@ -856,12 +856,12 @@ impl<'slice> CrnUnpacker<'slice>{
                         }
                         for c in 0..3 as usize{
                             if diff != 0 {
-                                block_endpoint[c] = (e0[c] << 3 | (((e1[c] as i32) - (e0[c] as i32)) & 7) as u8) as u8;
+                                block_endpoint[c] = (e0[c] << 3 | (((e1[c] as i32) - (e0[c] as i32)) & 7) as u8);
                             }else{
-                                block_endpoint[c] = ((e0[c] << 3 & 0xF0) | e1[c] >> 1) as u8;
+                                block_endpoint[c] = ((e0[c] << 3 & 0xF0) | e1[c] >> 1);
                             }
                         }
-                        block_endpoint[3] = (e0[3] << 5 | e1[3] << 2 | diff << 1 | flip) as u8;
+                        block_endpoint[3] = (e0[3] << 5 | e1[3] << 2 | diff << 1 | flip);
                         p_dst[data_pos * 4 + 0] = block_endpoint[0];
                         p_dst[data_pos * 4 + 1] = block_endpoint[1];
                         p_dst[data_pos * 4 + 2] = block_endpoint[2];
@@ -976,12 +976,12 @@ impl<'slice> CrnUnpacker<'slice>{
                         }
                         for c in 0..3 as usize{
                             if diff != 0 {
-                                block_endpoint[c] = (e0[c] << 3 | (((e1[c] as i32) - (e0[c] as i32)) & 7) as u8) as u8;
+                                block_endpoint[c] = (e0[c] << 3 | (((e1[c] as i32) - (e0[c] as i32)) & 7) as u8);
                             }else{
-                                block_endpoint[c] = ((e0[c] << 3 & 0xF0) | e1[c] >> 1) as u8;
+                                block_endpoint[c] = ((e0[c] << 3 & 0xF0) | e1[c] >> 1);
                             }
                         }
-                        block_endpoint[3] = (e0[3] << 5 | e1[3] << 2 | diff << 1 | flip) as u8;
+                        block_endpoint[3] = (e0[3] << 5 | e1[3] << 2 | diff << 1 | flip);
                         let p_alpha0_selectors: &[u16];
                         if flip != 0 {
                             p_alpha0_selectors = &self.m_alpha_selectors[alpha0_selector_index * 6 + 3..];
